Cambio 12: Crear, Leer, Actualizar y Eliminar (CRUD) completo por los usuarios

+++ routes/web.php
@@ -1,18 +1,42 @@
 <?php
 
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//CONTROLADORES:
 use App\Http\Controllers\HomeController;        //Login
+use App\Http\Controllers\LoginController;       //Controlador de Login
 use App\Http\Controllers\PostController;        //Controlador de Posts
 use App\Http\Controllers\AdminController;       //Controlador de Administradores
 use App\Http\Controllers\UserController;        //Controlador de Usuarios
 use App\Http\Controllers\FileController;        //Controlador de Archivos
 use App\Http\Controllers\AllUsersController;    //Controlador de Usuarios: Admnistradores/Usuarios
 use Illuminate\Support\Facades\Route;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//MODELOS:
 use App\Models\User;
 use App\Models\Post;
 
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //GLOBAL:
-Route::get('/login', function () {
-    return view('login');
+Route::get('/login', [HomeController::class, 'index']) -> name('login'); //Ruta para el login
+
+Route::post('/login', [LoginController::class, 'login']) ->name('login'); //Ruta para el login
+
+Route::post('/logout', [LoginController::class, 'logout'])->name('logout');
+
+// Rutas protegidas por autenticación
+Route::middleware('auth')->group(function () {
+    Route::get('/admin/administradores', function () {
+        return view('admin.dashboard');
+    })->name('pages.admin.admin');
+
+    Route::get('/admin/administradores', function () {
+        return view('editor.dashboard');
+    })->name('pages.admin.editor');
+
+    Route::get('/admin/administradores', function () {
+        return view('viewer.dashboard');
+    })->name('pages.admin.viewer');
 });
 
 Route::get('/prueba', function () {
@@ -24,7 +48,7 @@ Route::get('/admin/version', function () {
 });
 
 
-
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //PRUEBAS DE FRONTEND:
 
 //Route::get('/admin/perfil', [PostController::class, 'index']) -> name('pages.admin.perfil');
@@ -34,7 +58,7 @@ Route::get('/admin/version', function () {
 });*/
 
 
-
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //PRUEBAS DE BACKEND:
 Route::get('/', [HomeController::class, 'index']) -> name ('home');
 
@@ -61,13 +85,19 @@ Route::delete('/posts/{post}', [PostController::class, 'destroy']);
 //ADMINISTRADOR:
 Route::get('/admin/administradores', [AdminController::class, 'index']) -> name('pages.admin.admin');   //Muestra los datos de los admnistradores
 
-Route::post('/admin/administradores', [AdminController::class, 'created']) -> name('pages.admin.create');   //Crea un nuevo admnistrador
+Route::post('/admin/administradores', [AdminController::class, 'created']) -> name('pages.admin.admin.create');   //Crea un nuevo admnistrador
 
-Route::put('/admin/administradores/{admin}', [AdminController::class, 'update']) -> name('pages.admin.update');   //Actualiza los datos de los admnistradores
+Route::put('/admin/administradores/{admin}', [AdminController::class, 'update']) -> name('pages.admin.admin.update');   //Actualiza los datos de los admnistradores
 
-Route::delete('/admin/administradores/{admin}', [AdminController::class, 'delete']) -> name('pages.admin.delete');   //Muestra el formulario para eliminar un admnistrador
+Route::delete('/admin/administradores/{admin}', [AdminController::class, 'delete']) -> name('pages.admin.admin.delete');   //Muestra el formulario para eliminar un admnistrador
 
 Route::get('/admin/usuarios', [UserController::class, 'index']) -> name('pages.admin.users');           //Muestra los datos de los usuarios
+
+Route::post('/admin/usuarios', [UserController::class, 'created']) -> name('pages.admin.users.create');   //Crea un nuevo usuario
+
+Route::put('/admin/usuarios/{user}', [UserController::class, 'update']) -> name('pages.admin.users.update');   //Actualiza los datos de los usuarios
+
+Route::delete('/admin/usuarios/{user}', [UserController::class, 'delete']) -> name('pages.admin.users.delete');   //Muestra el formulario para eliminar un usuario
 
 Route::get('/admin/configuracion', [AdminController::class, 'config']) -> name('pages.admin.settings');     //Muestra la configuracion del administrador
